+++ ptv_timetable/__init__.py
@@ -3,10 +3,11 @@ from datetime import datetime, timezone
 from hashlib import sha1
 from hmac import HMAC
 from ratelimit import limits, sleep_and_retry
+from requests.models import Response
+from requests.sessions import Session
 from typing import Final, Literal, overload, Self, TypedDict
 import logging
 import platform
-import requests
 import urllib.parse
 if platform.system() == "Windows":
     import tzdata
@@ -30,7 +31,7 @@ _logger.addHandler(logging.NullHandler())
 class TimetableAPI(object):
     """Interface class with the PTV Timetable API."""
 
-    def __init__[**_P, _R](self: Self, dev_id: str | int, key: str, *, calls: int = 1, period: float = 10, ratelimit_handler: Callable[[Callable[_P, _R]], Callable[_P, _R]] = sleep_and_retry) -> None:
+    def __init__[**_P, _R](self: Self, dev_id: str | int, key: str, *, calls: int = 1, period: float = 10, ratelimit_handler: Callable[[Callable[_P, _R]], Callable[_P, _R]] = sleep_and_retry, session: Session | None = None) -> None:
         """Initialises a new :class:`TimetableAPI` instance with the supplied credentials.
 
         :param dev_id:            User ID
@@ -38,6 +39,7 @@ class TimetableAPI(object):
         :param calls:             Maximum number of calls that can be made to the API within the specified ``period``
         :param period:            Number of seconds since the last reset (or initialisation) at which the rate limiter will reset its call count
         :param ratelimit_handler: Function decorator that handles :class:`ratelimit.exception.RateLimitException` without re-raising it; defaults to :function:`ratelimit.decorators.sleep_and_retry`. A custom handler should match the specified signature, otherwise the program's behaviour is undefined (there is no runtime checking of the suitability of the handler)
+        :param session:           If specified, calls will be made using this HTTP session (default is to create a new :class:`requests.sessions.Session` instance)
         :return:                  ``None``
         """
 
@@ -53,20 +55,28 @@ class TimetableAPI(object):
         """API user ID"""
         self._key: Final[bytes] = key.encode(encoding="ascii")
         """API request signing key"""
-        self._get: Callable[..., requests.models.Response] = ratelimit_handler(limits(calls, period)(requests.get))
+
+        self._session = session if session is not None else Session()
+        """HTTP session used to make requests"""
+        self._is_user_session: Final[bool] = True if session is not None else False
+        """Whether the session is user-supplied (and therefore whether to auto-close on instance deletion)"""
+
+        self._get: Callable[..., Response] = ratelimit_handler(limits(calls, period)(self._session.get))
         """requests.get() function but rate-limited"""
 
         _logger.info("TimetableAPI instance created")
         return
 
     def __del__(self: Self) -> None:
-        """Logs the prospective deletion of an instance into the module logger once there are no more references to it in the program.
+        """Closes the underlying HTTP session if it wasn't supplied by the user and logs the prospective deletion of an instance into the module logger once there are no more references to it in the program.
 
         Note that Python does not guarantee that this will be called for any instance.
 
         :return: ``None``
         """
 
+        if not self._is_user_session:
+            self._session.close()
         _logger.info("TimetableAPI instance deleted")
         return
 
@@ -127,7 +137,7 @@ class TimetableAPI(object):
 
         url = self._encode_url(request)
         _logger.debug("Requesting from: " + url)
-        r: requests.models.Response = self._get(url)
+        r: Response = self._get(url)
         try:
             r.raise_for_status()
         except Exception: